Run scripts/programs in /etc/entrypoint.d
Sometimes it is useful to add scripts or binaries to an image which will be run
automatically when the container is launched but before it starts the main
daemon or service. These can be added by a descendent image or by being
bind-mounted into the container at runtime. This saves having to fork the whole
repo just make a very minor change to entry.sh. This is a rather common pattern
in some of the more popular images on Docker Hub. I feel it makes sense for a
fairly general-purpose image like this one as well.